pyproject.toml
[tool.poetry]
name = "st_experiments"
version = "0.1.0"
description = ""
authors = ["Leejay <leejayxia@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.6"
streamlit = "^0.82.0"
watchdog = "^2.1.2"
certifi = "^2020.12.5"
streamlit-aggrid = "^0.1.9"
st-flashcard = "^0.0.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
